+++ packages/sidecar/lib/src/server/starters/cli_starter.dart
@@ -7,6 +7,7 @@ import 'package:riverpod/riverpod.dart';
 import '../../client/cli_client.dart';
 import '../../client/client.dart';
 import '../../client/hot_reloader.dart';
+import '../../reports/reporter.dart';
 import '../../reports/stdout_reporter.dart';
 
 /// Run Sidecar from CLI or Debugger.
@@ -16,13 +17,15 @@ Future<void> startSidecarCli(
   bool isStrictMode = false,
 }) async {
   final isDebug = args.any((arg) => arg == 'debug' || arg == '--debug');
+  final _reporterProvider = stdoutReportProvider;
   final container = ProviderContainer(overrides: [
     analyzerClientProvider.overrideWithProvider(cliClientProvider),
+    reporterProvider.overrideWithProvider(_reporterProvider),
   ]);
   await runZonedGuarded<Future<void>>(
     () async {
       final client = container.read(analyzerClientProvider);
-      final reporter = container.read(stdoutReportProvider);
+      final reporter = container.read(reporterProvider);
       await client.openWorkspace();
       final hasErrors = reporter.hasErrors(isStrictMode: isStrictMode);
       reporter.print();