+++ packages/sidecar/lib/test.dart
@@ -1,4 +1,0 @@
-@Deprecated('Depend on package:sidecar_test instead')
-library sidecar.test;
-
-export 'package:sidecar_test/sidecar_test.dart';