+++ packages/sidecar_lints/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+- chore: Upgrade to Analyzer 6.0.0, latest Sidecar
+
 ## 0.1.0
 
 - new rule: rule_is_not_accessible
@@ -10,7 +14,7 @@
 
 ## 0.1.0-dev.3
 
-- fix: Remove old test lint from pubspec 
+- fix: Remove old test lint from pubspec
 
 ## 0.1.0-dev.2