+++ packages/sidecar_test/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.2
+
+- chore: Upgrade file dependency to latest
+
+## 0.1.1
+
+- chore: Upgrade Analyzer to v6.0.0
+
 ## 0.1.0
 
 - Initial release with basic utilities.